Introduce ArrayFieldBase in ra-core
